requirements.txt
# Core dependencies
pymongo>=4.6.0
redis>=5.0.0
pydantic>=2.5.0
pydantic[email]>=2.5.0
click>=8.1.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
mongomock>=4.1.2
fakeredis>=2.20.0

# Development
python-dotenv>=1.0.0
